Update cpu_power.csv
As described in issue #532, I merged the TDP values with the data in the eco2ai project. For contrasting values, a search was made to find the right one or the value from code carbon was kept; except when the difference was inferior to 1 (due to int-float difference), motivating the adoption of eco2ai's float value.
The list of contrasting values can be found in the issue.