Add ability to speficy GPUs by UUID prefix (#923)
* Add the ability to pass UUID prefixes as a way of specifying what GPUs to track..

This is to address https://github.com/mlco2/codecarbon/issues/873
Prior to this change you could only pass an index into the number of GPUs on the system.
Now you can pass a UUID prefix, including the 'MIG-' prefix per
https://docs.nvidia.com/cuda/cuda-c-programming-guide/index.html#cuda-environment-variables
if desired.

Note that I have not been able to test this on real life repo. The reporter of
https://github.com/mlco2/codecarbon/issues/873 was not able to provide a repro.

* Update logging and documentation for GPU indentifier prefix matching.

* Update the rst parameters file for GPU prefix matching and generate the matching html file.