Feature/add water usage (#941)
* Add support for setting WUE (water usage effectiveness) via a flag and calculating water usage.

This adds a --wue flag for a L/kWh WUE value.
I also fix some tests that for some reason started failing on Mac. Unclear why they are failing
now.

* Create documentation and an example script for WUE.

Also remove an unecessary asserts in unit.py that was breaking the pue.py example script. It's uncessary because we only need to assert one number in a product is a float to get a float result.

* Check scheduler exists before checking if it has stopped - a cause of test flakiness